# ai-specs/specs/api-spec.yml
openapi: 3.1.0
info:
  title: ADA ASM API
  version: 0.1.0
  description: |
    REST API for ADA ASM. This file is the source of truth for the API
    contract and is regenerated from the FastAPI app's `/openapi.json`
    on every PR that changes `backend/app/api/**`.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/supplier-sync/runs:
    get:
      tags: [supplier-sync]
      summary: List recent supplier sync runs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: supplier
          schema:
            type: string
            enum: [mouser, digikey, tme, farnell, rs]
      responses:
        "200":
          description: Paginated list of runs, ordered by `started_at DESC`.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SupplierSyncRunResponse" }
        "401": { description: Unauthenticated. }
    post:
      tags: [supplier-sync]
      summary: Enqueue an ad-hoc supplier sync (no waiting for the daily Beat)
      description: >
        Pre-creates the `supplier_sync_runs` row HTTP-side so the
        operator gets the real `run_id` back synchronously, then
        enqueues `sync_one_supplier.delay(code, run_id)` on Celery. Poll
        `GET /api/v1/supplier-sync/runs/{run_id}` for completion. Fails
        with 422 `SUPPLIER_NOT_ENABLED` when the supplier is not listed
        in `SUPPLIER_SYNC_ENABLED_SUPPLIERS` or has no credentials.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: supplier
          required: true
          schema:
            type: string
            enum: [mouser, digikey, tme, farnell, rs]
      responses:
        "202":
          description: Sync task enqueued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TriggerSyncResponse" }
        "401": { description: Unauthenticated. }
        "422":
          description: Supplier not enabled or has no credentials.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  code: { type: string, enum: [SUPPLIER_NOT_ENABLED] }
  /api/v1/supplier-sync/runs/{run_id}/errors:
    get:
      tags: [supplier-sync]
      summary: Per-component errors captured during one sync run
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: run_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        "200":
          description: Errors for the run, ordered by `occurred_at DESC`.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SupplierSyncErrorResponse" }
        "404":
          description: Unknown run id.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  code: { type: string, enum: [SUPPLIER_SYNC_RUN_NOT_FOUND] }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SupplierCode:
      type: string
      enum: [mouser, digikey, tme, farnell, rs]
...
    SyncRunStatus:
      type: string
      enum: [running, success, partial, failed]
    SyncErrorCode:
      type: string
      enum:
        - RATE_LIMITED
        - NOT_FOUND
        - HTTP_5XX
        - PARSE_ERROR
        - AUTH_FAILED
        - FX_UNAVAILABLE
        - TIMEOUT
        - UNKNOWN
    SupplierSyncRunResponse:
      type: object
      required: [id, supplier, started_at, components_processed, components_updated, errors_count, status]
      properties:
        id: { type: string, format: uuid }
        supplier: { $ref: "#/components/schemas/SupplierCode" }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        components_processed: { type: integer, minimum: 0 }
        components_updated: { type: integer, minimum: 0 }
        errors_count: { type: integer, minimum: 0 }
        status: { $ref: "#/components/schemas/SyncRunStatus" }
        error_summary: { type: string, nullable: true }
    SupplierSyncErrorResponse:
      type: object
      required: [id, run_id, component_id, supplier, error_code, error_message, occurred_at]
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        component_id: { type: string, format: uuid }
        supplier: { $ref: "#/components/schemas/SupplierCode" }
        error_code: { $ref: "#/components/schemas/SyncErrorCode" }
        error_message: { type: string }
        occurred_at: { type: string, format: date-time }
    TriggerSyncResponse:
      type: object
      required: [run_id, task_id]
      properties:
        run_id:
          type: string
          format: uuid
          description: ID of the `supplier_sync_runs` row created.
        task_id:
          type: string
          description: Celery task ID for stronger correlation.
...